--- a/final_grade_a_results.json
+++ b/final_grade_a_results.json
@@ -1,0 +1,6 @@
+{
+  "timestamp": "2025-09-10T12:17:27.700783",
+  "final_score": 89.22813036020582,
+  "grade": "B",
+  "total_breeds": 583,
+  "breeds_with_weight":